pyproject.toml
[tool.poetry]
name = "programming-for-biology"
version = "0.1.0"
description = ""
authors = ["Jan Nils Ferner <contact@jnferner.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
numpy = "^1.21.2"
scipy = "^1.7.1"
matplotlib = "^3.4.3"
biopython = "^1.79"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
